refactor: place Snake and Fruit in separate files
